Added integration test
This is a test yaml file that includes all of the currently planned
features, as well as a header file that the tests are designed to test.
I plan to add a C file that actually implements the header once I reach
the point that I care about testing that the generated output is
correct.