More "generated" code
This is the meat of how I imagine the generated code will look like.
I've come to a couple of conclusions: thus far, my yaml syntax seems
like it will be reasonably able to handle the code I'll need to write.
This is good, it means that my effort in planning out the language was
not in vain. The other thing I noticed, though, was that I don't think
that type inference will work as well as I hoped. I think I'm going to
have to have some sort of type annotation to make this work correctly.