Fixed render_main to pass its test
Specifically: Fixed the arguments passed to template.render and made
ostester a package as required by PackageLoader
It still doesn't work, though. A demonstration of the limitations of
unit tests in catching bugs. A functional test where I take the output
and feed it into a C compiler would have caught this particular bug,
but I'm not at that point yet.