Added test for invalid header
I decided to disallow headers that don't end in .h. There's room to
disagree with my decision, but I'm already making impositions on the
form of the C code this tests, and I don't think that this one is
particularly burdensome. I'm considering adding support for header files
ending in .c, but that's a decision for another time.