Added another test for invalid header names
In strict TDD, this is disallowed, since the test didn't fail. But I'm
doing it anyway as a way of specifying the behavior of the function in
an area where it doesn't do what the user would expect. As to why it has
that behavior? Part laziness, part I don't think it's much of an
imposition to enforce that names are alphanumeric and it's easily
relaxed to allow other characters in names.