Fixed bugs exposed by the suite unit test
It's nice having some validation that your tests are working right. Of
course, this is the reason that, in TDD proper, you're supposed to only
write failing tests. In this case, though, I wrote the templates that
I'm testing first to minimize the risk of rewriting significant swaths
of Python code as I learn what exactly I need in templates. Perhaps I
could have designed the templates better to satisfy both of these goals?