Added tests that Sig's inputs + output are CTypes
This is an example of a corrolary to the idea that deserialiation
functions are basically constructors and should be treated as such.
Running them through the constructor is definitely the best way to be
sure that they don't have subtly different class invariants, but make
sure that the constructor is the one actually enforcing the class
invariants.
It isn't, so these tests fail.