Required compare_with function of BinOp
Replaces c_code.
This change is basically me exploring the API design space. c_code on a
bunch of different functions would have been a reasonable way of
designing the API, I'm just trying something different.
Of course, compare_with has the disadvantage of suggesting that it
compares two things, rather than generates code that compares two
things. That's something to keep in mind as I continue working on this