Added __main__ to the ostester package
This allows loading tests with python -m ostester -t, with an optional
-v flag for verbose outputs - this includes the generated C code.
The -m flag is both required and annoying. I'm going to look into ways
to make it non-required, but for now it's needed to run the test suite.

This change also creates the framework for adding a command line
interface, which will be important when it comes time to actually use
this in the operating system