Added factory for C Types
It has become apparant that my current strategy for dealing with types
is not able to handle properly initializing values in C. So, I'm making
CType a factory that delegates to a more specialized class.
Naturally, this change breaks one of my tests... I'm undecided whether
this means that my tests were too brittle or not.