Renamed SimpleCType to LiteralCType
After realizing that char*s have literals as well. This is going to
require a rethink of my design, and it might result in no change -
strings always being initialized as char arrays, rather than using
string literals.
One of the perks of writing a code generater, rather than writing the
generated code by hand, is the ability to ignore convenience features
such as string literals.
That said, it would be nice to incorporate them, if only to make reading
the generated code a little easier.