Begun adding a combo value + type class
I need the ability to initialize variables ahead of time, then use their
names in functions and other initializations. This ability needs to be
stuck somewhere, and there's currently no good place for it.
I can't say I'm thrilled with the name, though. That might change at
some point in the future.