First commit towards a type-inferred value
Doing the AST transform, I realized that TypeValue didn't satisfy the
actual use-case I needed. Because I'm hoping that most values in the
YAML files will be type-inferred, a class that requires a type to
initialize a value is a non-starter. Rather, I'm thinking I need a class
that can take a value and coerce it to the actual type needed.