Reworked SimpleCType
In preparation for reworking how I implement types. This change has been
waiting in the wings for a while, CType isn't supposed to have any code
to handle arrays or pointers, and that's what I removed.
One thing that bugs me, though. I had c_type raising a
NotImplementedError when faced with a pointer or array type, but that
broke a couple of tests. What bugs me, though, is that I removed the
raise, and now all my tests pass. Convenient from the standpoint of
having my tests pass, but annoying because I made a change that broke
code, and my tests don't catch it. Sigh.