Added more detail to test_new_declarations
Also fixed a bug with recursive_declarations. It now works as I intend
it to, although I'm not sure that how I intend it to is how it will end
up working.
I'm also unhappy with the code in test_new_declarations. It could use
improvement, but that can wait until the next time I touch the code